Fixes to make video playback smoother
Changed reported blksize to 128K which made a big difference to
playback.

Prevented LoggingMixIn from logging return values for reads because they
made playback laggy when debug output was enabled and the logs
unreadable because everything was logged. Now only the length of bytes
is written.

Newer recordings still don't work and even their XML files are not
downloaded correctly.